WFD Implementation: Fix for IPv4
When a node got disconnected from the Access Point, its DHCP assigned IP
for an interface becomes unspesified. Some other nodes may still think
that the node is still connected and send TCP segments. The frame will
eventually forwarded up by the mac layer because its mac address is
correct. The network layer will pass the segment up to TCP. TCP will
reject it and send it as a negative ACK, by reversing the src and dst ip
addressed in the segment. The netowrk layer will recieve the packet and
panic in such case because the src ip address does not match any
interface in the routing table.
My fix for now is to ignore such packet.